requirements.txt
einops
tqdm
transformers
matplotlib
pyyaml
scipy